Move seed logic to service layer
